==> 2_MergeDataSet.ipynb <==
import pandas as pd
import os

# 결과를 저장할 빈 데이터프레임 생성
merged_df = pd.DataFrame()

dfs = []

# 각각의 엑셀 파일에서 데이터를 읽어와 통합
file_prefix = ["국가애도기간", "용어지침", "대통령사과", "경찰책임", "정부책임", "유가족협의회", "특별법"]

for prefix in file_prefix:
    print(f"{prefix}파일 병합중...")
    file_size = 0
    merged_df = pd.DataFrame()
    i = 1
    while True:
        file_path = f"./1_crawled/{prefix}{i}.xlsx"
        if os.path.exists(file_path):
            df = pd.read_excel(file_path)
            print(f"\t{prefix}{i}.xlsx = {df.shape[0]}개")
            file_size += df.shape[0]
            merged_df = pd.concat([merged_df, df], ignore_index=True)
            i += 1
        else:
            dfs.append(merged_df)
            print(f"#병합끝 : 총{file_size}개 데이터")
            break

dfs[0]

dfs[1]

# 결과를 저장할 디렉토리 생성
output_directory = "./2_raw_dataset/"
os.makedirs(output_directory, exist_ok=True)

# 각각의 통합된 데이터프레임을 Excel 파일로 저장
for prefix, df in zip(file_prefix, dfs):
    output_file_path = f"{output_directory}{prefix}_merged.xlsx"
    df.to_excel(output_file_path, index=False)
    print(f"{output_file_path}에 저장 완료")
