==> 3_dataPreprocess.ipynb <==
import os
import pandas as pd
import re

# 파일 접두어 정의
file_prefixes = ["국가애도기간", "용어지침", "대통령사과", "경찰책임", "정부책임", "유가족협의회", "특별법"]

# 디렉토리 경로
directory_path = "./2_raw_dataset"

dfs=[]

# 각 파일 접두어에 대해 반복
for prefix in file_prefixes:
    # 파일 경로 정의
    file_path = os.path.join(directory_path, f"{prefix}_merged.xlsx")

    # 파일이 존재하는지 확인
    if os.path.exists(file_path):
        df = pd.read_excel(file_path)
        
        dfs.append(df)
    else:
        print(f"{prefix}_merged.xlsx 파일을 찾을 수 없습니다.")

dfs[0]

print(dfs[0]['Likes'].dtype)

dfs[1]

print(dfs[1]['Likes'].dtype)

# 날짜를 추출하는 함수 정의
def extract_date(text):
    match = re.search(r'\d{4}\. \d{2}\. \d{2}', text)
    if match:
        return match.group()
    else:
        return text

for df in dfs:
    df['Times'] = df['Times'].apply(extract_date)

dfs[0]

# Likes 컬럼의 값을 숫자로 변환하는 함수 정의
def convert_likes(likes_str):
    if isinstance(likes_str, str):  # 문자열인 경우에만 처리
        if '천' in likes_str:
            return float(likes_str.replace('천', '')) * 1000
        elif '만' in likes_str:
            return float(likes_str.replace('만', '')) * 10000
        else:
            return float(likes_str)
    else:
        return likes_str  # 정수인 경우에는 그대로 반환

# 각 데이터프레임에 대해 작업 수행
for df in dfs:
    # Likes 컬럼의 데이터 타입에 따라 처리
    if df['Likes'].dtype != 'float64':
        df['Likes'] = df['Likes'].apply(convert_likes)

for i,df in enumerate(dfs):
    print(f"{file_prefixes[i]} 파일 null처리중...")
    print(f"\t처리전 null 개수: {df['Comment'].isnull().sum()}")
    df.dropna(subset=['Comment'], inplace=True)
    print(f"\t처리후 null 개수: {df['Comment'].isnull().sum()}")

for i,df in enumerate(dfs):
    print(f"{file_prefixes[i]} 파일 도배 댓글 처리중...")
    print(f"\t처리전 행 개수: {df.shape[0]}")
    df.drop_duplicates(subset=['Username', 'Comment'], inplace=True)
    print(f"\t처리후 행 개수: {df.shape[0]}")

# 결과를 저장할 디렉토리 생성
output_directory = "./3_processed_dataset/"
os.makedirs(output_directory, exist_ok=True)

# 각각의 통합된 데이터프레임을 Excel 파일로 저장
for prefix, df in zip(file_prefixes, dfs):
    output_file_path = f"{output_directory}{prefix}_processed.xlsx"
    df.to_excel(output_file_path, index=False)
    print(f"{output_file_path}에 저장 완료")
