==> 4_trainDataProcess.ipynb <==
# 오픈소스 학습 데이터셋 가져오기

import pandas as pd

## nsmc
- 긍정(1), 부정(0) 데이터 셋

file1 = './train/긍부정데이터셋.txt'

data1 = pd.read_table(file1, index_col='id')

data1

# 라벨 종류 확인
unique_label1 = data1['label'].unique()

print(unique_label1)

data1.to_excel('./train/긍부정데이터셋.xlsx', engine='openpyxl')

## 2번쨰 데이터 셋
- '공포' '놀람' '분노' '슬픔' '중립' '행복' '혐오' -> 7가지 감정 데이터 셋

file2 = './train/7가지감정데이터셋.xlsx'

data2 = pd.read_excel(file2, usecols=['Sentence', 'Emotion'])

data2

# 라벨 종류 확인
unique_label2 = data2['Emotion'].unique()

print(unique_label2)

## 3번째 데이터 셋
- 혐오 데이터 셋

file3 = './train/혐오단어데이터셋.tsv'

data3 = pd.read_csv(file3, delimiter='\t')

data3

# 라벨 종류 확인
unique_label3 = data3['hate'].unique()

print(unique_label3)

data3.to_excel('./train/혐오단어데이터셋.xlsx', engine='openpyxl')
