==> 7_topicExtraction.ipynb <==
!pip install konlpy

import pandas as pd
from konlpy.tag import Okt
okt = Okt()
from konlpy.tag import Kkma
kkma = Kkma()
from collections import Counter
from google.colab import drive
import os

# 구글 드라이브 마운트
drive.mount('/content/gdrive')

# 파일 접두어 정의
file_prefixes = ["국가애도기간", "대통령사과", "경찰책임", "정부책임", "유가족협의회", "특별법"]

# 디렉토리 경로
directory_path = "/content/gdrive/MyDrive/6_emotion/"

dfs=[]

# 각 파일 접두어에 대해 반복
for prefix in file_prefixes:
    # 파일 경로 정의
    file_path = directory_path + f"{prefix}_emotion.xlsx"

    df = pd.read_excel(file_path)

    dfs.append(df)

def extract_and_merge_nouns(text):
    # 명사 추출
    nouns = okt.nouns(text)

    if len(nouns) == 0:
      return []

    # 연속된 명사들을 하나로 합치기
    merged_nouns = [nouns[0]]
    i = 1
    j = 0
    while i < len(nouns):
        now_noun = merged_nouns[j]
        next_noun = nouns[i]

        now_index = text.find(now_noun)
        now_length = len(now_noun)
        next_index = text.find(next_noun)

        if next_index == now_index + now_length:
          merged_nouns[j] = now_noun + next_noun
        else:
          merged_nouns.append(next_noun)
          j += 1

        i += 1

    #크기가 1인 명사 제거
    merged_nouns = [noun for noun in merged_nouns if len(noun) > 1]
    #merged_nouns = [noun for noun in nouns if len(noun) > 1]
    return merged_nouns


desired_pos_tags = ['NNG', 'NNP']

def extract_and_merge_nouns(text):
    try:
      # 명사 추출
      nouns_list = kkma.pos(text)

      filtered_nouns_list = []

      for doc in nouns_list:
          if doc[1] in desired_pos_tags:
            filtered_nouns_list.append(doc[0])


      #크기가 1인 명사 제거
      merged_nouns = [noun for noun in filtered_nouns_list if len(noun) > 1]

      return merged_nouns

    except Exception as e:
        return []



for df in dfs:
    print(f'명사 추출 진행중...')
    df['nouns_list'] = df['Comment'].apply(extract_and_merge_nouns)

dfs[0]

emotion_class = ['슬픔', '분노', '행복', '혐오', '중립', '놀람', '공포']

result_dfs = []

for df in dfs:
  # 첫번째 작업: Likes수만큼 nouns_list를 중복시켜서 copy_df에 저장
  copy_df = df.loc[df.index.repeat(df['Likes'].fillna(1))]

  # 두번째 작업: copy_df의 nouns_list에서 가장 많이 등장한 명사 10개 추출(top_nouns)
  nouns_counter = Counter([noun for nouns_list in copy_df['nouns_list'] for noun in nouns_list])
  top_nouns = [noun for noun, _ in nouns_counter.most_common(10)]
  print(nouns_counter.most_common(10))

  # 세번째 작업: top_nouns명사들이 각 Emotion에서 몇번 등장하였는지 아래 형태로 result_df 생성
  result_df = pd.DataFrame(index=top_nouns, columns=copy_df['Emotion'].unique()).fillna(0)
  for _, row in copy_df.iterrows():
      for noun in row['nouns_list']:
          if noun in top_nouns:
              result_df.loc[noun, row['Emotion']] += 1
  result_dfs.append(result_df)

result_dfs[0]

# 결과를 저장할 디렉토리 생성
output_directory = "/content/gdrive/MyDrive/7_topic_3/"
os.makedirs(output_directory, exist_ok=True)

for i, result_df in enumerate(result_dfs):
    prefix = file_prefixes[i]
    output_file_path = f"{output_directory}{prefix}_topic.xlsx"
    result_df.to_excel(output_file_path, index=True)
    print(f"{output_file_path}에 저장 완료")
